Refactor cacheing objects into their classes
